Removed flag and made the methods always sort. Moved tests to model
