+++ cellmlmanip/model.py
@@ -495,29 +495,19 @@ class Model(object):
 
         return eqs
 
-    def get_derivative_symbols(self, order_by_order_added=False):
+    def get_derivative_symbols(self):
         """Returns a list of derivative symbols found in the given model graph.
-
-        :param order_by_order_added: indicates whether state_symbols are sorted in the order they appear in the model.
-        Please Note: the order of returned derivative symbols is not guaranteed
-                     specificly the order is not reproducible on Python <= 3.5 without the flag
+           The list is ordered by appearance in the cellml document.
         """
         derivative_symbols = [v for v in self.graph if isinstance(v, sympy.Derivative)]
-        if order_by_order_added:
-            return sorted(derivative_symbols, key=lambda state_var: self.get_meta_dummy(state_var.args[0]).order_added)
-        return derivative_symbols
+        return sorted(derivative_symbols, key=lambda state_var: self.get_meta_dummy(state_var.args[0]).order_added)
 
     def get_state_symbols(self, order_by_order_added=False):
         """Returns a list of state variables found in the given model graph.
-
-        :param order_by_order_added: indicates whether state_symbols are sorted in the order they appear in the model.
-        Please Note: the order of returned derivative symbols is not guaranteed
-                     specificly the order is not reproducible on Python <= 3.5 without the flag
+           The list is ordered by appearance in the cellml document.
         """
         state_symbols = [v.args[0] for v in self.get_derivative_symbols()]
-        if order_by_order_added:
-            return sorted(state_symbols, key=lambda state_var: self.get_meta_dummy(state_var).order_added)
-        return state_symbols
+        return sorted(state_symbols, key=lambda state_var: self.get_meta_dummy(state_var).order_added)
 
     def get_free_variable_symbol(self):
         """Returns the free variable of the given model graph.
