+++ tests/test_aslanidi_model.py
@@ -27,3 +27,64 @@ class TestAslanidiModel:
         model.get_equation_graph()
         membrane_voltage = model.get_symbol_by_ontology_term(OXMETA, "membrane_voltage")
         assert(model.get_initial_value(membrane_voltage) == -80)
+
+    def test_state_symbols(self, model):
+        """
+        _membrane$V
+        _sodium_channel_m_gate$m
+        _sodium_channel_h_gate$h
+        _potassium_channel_n_gate$n
+        """
+        state_symbols = model.get_state_symbols()
+        state_symbols_ordered_by_input = model.get_state_symbols(order_by_order_added=True)
+        assert len(state_symbols) == len(state_symbols_ordered_by_input) == 29
+        assert set(state_symbols) == set(state_symbols_ordered_by_input)
+        assert str(state_symbols_ordered_by_input) == \
+            '[_membrane$V, _sodium_current_m_gate$m, _sodium_current_h1_gate$h1, _sodium_current_h2_gate$h2, '\
+            '_L_type_Ca_channel_d_L_gate$d_L, _L_type_Ca_channel_f_L_gate$f_L, '\
+            '_T_type_Ca_channel_d_T_gate$d_T, _T_type_Ca_channel_f_T_gate$f_T, '\
+            '_Ca_independent_transient_outward_K_current_r_gate$r, '\
+            '_Ca_independent_transient_outward_K_current_s1_gate$s1, '\
+            '_Ca_independent_transient_outward_K_current_s2_gate$s2, '\
+            '_Ca_independent_transient_outward_K_current_s3_gate$s3, _delayed_rectifier_K_current_z_gate$z, '\
+            '_delayed_rectifier_K_current_pa_gate$p_a, _delayed_rectifier_K_current_pi_gate$p_i, '\
+            '_intracellular_ion_concentrations$Na_i, _intracellular_ion_concentrations$Ca_i, '\
+            '_intracellular_ion_concentrations$K_i, _intracellular_Ca_buffering$O_C, '\
+            '_intracellular_Ca_buffering$O_TC, _intracellular_Ca_buffering$O_TMgC, '\
+            '_intracellular_Ca_buffering$O_TMgMg, _cleft_space_ion_concentrations$K_c, '\
+            '_Ca_handling_by_the_SR$Ca_rel, _Ca_handling_by_the_SR$Ca_up, _Ca_handling_by_the_SR$O_Calse, '\
+            '_Ca_handling_by_the_SR$F1, _Ca_handling_by_the_SR$F2, _Ca_handling_by_the_SR$F3]'
+
+    def test_derivative_symbols(self, model):
+        derivs = model.get_derivative_symbols()
+        assert len(derivs) == 29
+
+        assert str(derivs) == '[Derivative(_membrane$V, _environment$time), '\
+            'Derivative(_sodium_current_m_gate$m, _environment$time), '\
+            'Derivative(_sodium_current_h1_gate$h1, _environment$time), '\
+            'Derivative(_sodium_current_h2_gate$h2, _environment$time), '\
+            'Derivative(_L_type_Ca_channel_d_L_gate$d_L, _environment$time), '\
+            'Derivative(_L_type_Ca_channel_f_L_gate$f_L, _environment$time), '\
+            'Derivative(_T_type_Ca_channel_d_T_gate$d_T, _environment$time), '\
+            'Derivative(_T_type_Ca_channel_f_T_gate$f_T, _environment$time), '\
+            'Derivative(_Ca_independent_transient_outward_K_current_r_gate$r, _environment$time), '\
+            'Derivative(_Ca_independent_transient_outward_K_current_s1_gate$s1, _environment$time), '\
+            'Derivative(_Ca_independent_transient_outward_K_current_s2_gate$s2, _environment$time), '\
+            'Derivative(_Ca_independent_transient_outward_K_current_s3_gate$s3, _environment$time), '\
+            'Derivative(_delayed_rectifier_K_current_z_gate$z, _environment$time), '\
+            'Derivative(_delayed_rectifier_K_current_pa_gate$p_a, _environment$time), '\
+            'Derivative(_delayed_rectifier_K_current_pi_gate$p_i, _environment$time), '\
+            'Derivative(_intracellular_ion_concentrations$Na_i, _environment$time), '\
+            'Derivative(_intracellular_ion_concentrations$Ca_i, _environment$time), '\
+            'Derivative(_intracellular_ion_concentrations$K_i, _environment$time), '\
+            'Derivative(_intracellular_Ca_buffering$O_C, _environment$time), '\
+            'Derivative(_intracellular_Ca_buffering$O_TC, _environment$time), '\
+            'Derivative(_intracellular_Ca_buffering$O_TMgC, _environment$time), '\
+            'Derivative(_intracellular_Ca_buffering$O_TMgMg, _environment$time), '\
+            'Derivative(_cleft_space_ion_concentrations$K_c, _environment$time), '\
+            'Derivative(_Ca_handling_by_the_SR$Ca_rel, _environment$time), '\
+            'Derivative(_Ca_handling_by_the_SR$Ca_up, _environment$time), '\
+            'Derivative(_Ca_handling_by_the_SR$O_Calse, _environment$time), '\
+            'Derivative(_Ca_handling_by_the_SR$F1, _environment$time), '\
+            'Derivative(_Ca_handling_by_the_SR$F2, _environment$time), '\
+            'Derivative(_Ca_handling_by_the_SR$F3, _environment$time)]'
